This is a massive commit! It defines all the undefined variables and indexes in main index.php and the admin area. It fixes the following bugs : Database content for language x was being inputed into the database as scramble. We now specify that mysql uses UTF-8 and it prevents it. People using other languages can now read what is in their database. The contact module's session start was broken. Now fixed. The admin carousel is now fixed. (Thanks for the report Scott!) Forms now use UTF-8 for encoding and the title is below the meta. (Suggested by Isa!) Lots of other little fixes made as found along the way.
I have tested this as much as possible so far. If you find a problem, report it on the forum please or contact me by email,
failing that, put it on the bug tracker please. The nature of this commit means bugs are highly likely.
The installer has not had the same treatment yet but it may well do. I have not test installed this yet.
No new features added.